+++ src/musicscan/data/analyzer.py
@@ -26,9 +26,11 @@
 Testing code to identify possible edit flags.
 '''
 
+# pylint: disable=too-many-public-methods
+
 import re
 from musicscan.data.cd import Album
-from musicscan.data.flags import FlagCodes
+from musicscan.data.flags import AlbumFlagCodes, TrackFlagCodes
 from musicscan.data.track import Track
 
 
@@ -42,11 +44,25 @@ class Analyzer():
         '''
         Main album testing routine.
         '''
+        self.album_tests(in_album)
+        self.track_tests(in_album)
+        self.album_closing_tests(in_album)
+
+    def album_tests(self, in_album: Album):
+        '''
+        Perform tests against the album.
+        '''
         self.check_album_values(in_album)
         self.test_album_title(in_album)
         self.test_album_artist(in_album)
         self.test_soundtrack(in_album)
         self.test_score(in_album)
+        self.test_tribute(in_album)
+
+    def track_tests(self, in_album: Album):
+        '''
+        Perform tests against every track.
+        '''
         for dsc in sorted(in_album.discs):
             for trk in sorted(in_album.discs[dsc].tracks,
                               key=lambda x: x.track_no):
@@ -57,6 +73,15 @@ class Analyzer():
                 self.test_track_genre(trk)
                 self.test_album_track_blank(trk)
                 self.test_album_track_bonus(trk)
+                self.test_album_track_intro(trk)
+                self.test_album_track_reprise(trk)
+
+    def album_closing_tests(self, in_album: Album):
+        '''
+        Perform additional tests after all
+        tracks are analzed.
+        '''
+        self.test_intro_reprise_presence(in_album)
 
     def check_album_values(self, in_album: Album):
         '''
@@ -65,7 +90,7 @@ class Analyzer():
         '''
         if not in_album.title:
             in_album.title = 'PLACEHOLDER ALBUM TITLE'
-            in_album.flags.add_flag(FlagCodes.m_album_title)
+            in_album.flags.add_flag(TrackFlagCodes.m_album_title)
 
     def check_track_values(self, in_track: Track):
         '''
@@ -74,7 +99,7 @@ class Analyzer():
         '''
         if not in_track.artist:
             in_track.artist = 'PLACEHOLDER TRACK ARTIST'
-            in_track.flags.add_flag(FlagCodes.m_track_artist)
+            in_track.flags.add_flag(TrackFlagCodes.m_track_artist)
 
     def test_album_title(self, in_album: Album):
         '''
@@ -84,9 +109,9 @@ class Analyzer():
         greatest_p = re.compile(r'Greatest', re.IGNORECASE)
         hits_p = re.compile(r'\s+Hit', re.IGNORECASE)
         if greatest_p.search(in_album.title):
-            in_album.flags.add_flag(FlagCodes.p_greatest)
+            in_album.flags.add_flag(AlbumFlagCodes.p_greatest)
         if hits_p.search(in_album.title):
-            in_album.flags.add_flag(FlagCodes.p_hits_compo)
+            in_album.flags.add_flag(AlbumFlagCodes.p_hits_compo)
 
     def test_soundtrack(self, in_album: Album):
         '''
@@ -96,10 +121,10 @@ class Analyzer():
         first_track = in_album.first_track()
         soundtrack_p = re.compile(r'\s+Soundtrack\s+', re.IGNORECASE)
         if soundtrack_p.search(in_album.title):
-            in_album.flags.add_flag(FlagCodes.p_soundtrack)
+            in_album.flags.add_flag(AlbumFlagCodes.p_soundtrack)
         if first_track.genre == 'Soundtrack':
-            if FlagCodes.p_soundtrack not in in_album.flags.flags:
-                in_album.flags.add_flag(FlagCodes.p_soundtrack)
+            if AlbumFlagCodes.p_soundtrack not in in_album.flags.flags:
+                in_album.flags.add_flag(AlbumFlagCodes.p_soundtrack)
 
     def test_score(self, in_album: Album):
         '''
@@ -108,7 +133,16 @@ class Analyzer():
         '''
         score_p = re.compile(r'\s+Score\s+', re.IGNORECASE)
         if score_p.search(in_album.title):
-            in_album.flags.add_flag(FlagCodes.p_score)
+            in_album.flags.add_flag(AlbumFlagCodes.p_score)
+
+    def test_tribute(self, in_album: Album):
+        '''
+        Test if the album title suggests that it is a
+        tribute to another artist.
+        '''
+        tribute_p = re.compile(r'\s+Tribute\s+', re.IGNORECASE)
+        if tribute_p.search(in_album.title):
+            in_album.flags.add_flag(AlbumFlagCodes.p_tribute)
 
     def test_album_artist(self, in_album: Album):
         '''
@@ -117,7 +151,7 @@ class Analyzer():
         '''
         and_p = re.compile(r'\s+and\s+', re.IGNORECASE)
         if and_p.search(str(in_album.artist)):
-            in_album.flags.add_flag(FlagCodes.l_group)
+            in_album.flags.add_flag(TrackFlagCodes.l_group)
 
     def test_album_track_chrs(self, in_track: Track):
         '''
@@ -125,9 +159,9 @@ class Analyzer():
         have been embeded in the track name in parenthesis.
         '''
         if '[' in in_track.title or ']' in in_track.title:
-            in_track.flags.add_flag(FlagCodes.d_sq_brackets)
+            in_track.flags.add_flag(TrackFlagCodes.d_sq_brackets)
         if '(' in in_track.title or ')' in in_track.title:
-            in_track.flags.add_flag(FlagCodes.d_parenthesis)
+            in_track.flags.add_flag(TrackFlagCodes.d_parenthesis)
 
     def test_album_track_featured(self, in_track: Track):
         '''
@@ -137,9 +171,9 @@ class Analyzer():
         feature_p = re.compile(r'Feat', re.IGNORECASE)
         with_p = re.compile(r'With\s+', re.IGNORECASE)
         if feature_p.search(in_track.title) or with_p.search(in_track.title):
-            in_track.flags.add_flag(FlagCodes.p_feat_artist)
+            in_track.flags.add_flag(TrackFlagCodes.p_feat_artist)
         if feature_p.search(in_track.artist) or with_p.search(in_track.artist):
-            in_track.flags.add_flag(FlagCodes.p_feat_artist)
+            in_track.flags.add_flag(TrackFlagCodes.p_feat_artist)
 
     def test_album_track_live(self, in_track: Track):
         '''
@@ -148,7 +182,7 @@ class Analyzer():
         '''
         live_p = re.compile(r'Live\s+', re.IGNORECASE)
         if live_p.search(in_track.title):
-            in_track.flags.add_flag(FlagCodes.p_live)
+            in_track.flags.add_flag(TrackFlagCodes.p_live)
 
     def test_track_genre(self, in_track: Track):
         '''
@@ -157,7 +191,7 @@ class Analyzer():
         country_p = re.compile(r'Country & Folk', re.IGNORECASE)
         if in_track.genre:
             if country_p.search(in_track.genre):
-                in_track.flags.add_flag(FlagCodes.p_genre_country_folk)
+                in_track.flags.add_flag(TrackFlagCodes.p_genre_country_folk)
 
     def test_album_track_demo(self, in_track: Track):
         '''
@@ -166,7 +200,7 @@ class Analyzer():
         '''
         demo_p = re.compile(r'Demo', re.IGNORECASE)
         if demo_p.search(in_track.title):
-            in_track.flags.add_flag(FlagCodes.p_demo)
+            in_track.flags.add_flag(TrackFlagCodes.p_demo)
 
     def test_album_track_blank(self, in_track: Track):
         '''
@@ -175,7 +209,7 @@ class Analyzer():
         '''
         blank_p = re.compile(r'Blank', re.IGNORECASE)
         if blank_p.search(in_track.title):
-            in_track.flags.add_flag(FlagCodes.p_blank_track)
+            in_track.flags.add_flag(TrackFlagCodes.p_blank_track)
 
     def test_album_track_bonus(self, in_track: Track):
         '''
@@ -185,4 +219,42 @@ class Analyzer():
         '''
         bonus_p = re.compile(r'\W+Bonus', re.IGNORECASE)
         if bonus_p.search(in_track.title):
-            in_track.flags.add_flag(FlagCodes.p_bonus_track)
+            in_track.flags.add_flag(TrackFlagCodes.p_bonus_track)
+
+    def test_album_track_intro(self, in_track: Track):
+        '''
+        Test the track to see if it is an introduction to
+        another track, or possibly for the album itself.
+        '''
+        intro_p = re.compile(r'[\s\W]?Intro', re.IGNORECASE)
+        if intro_p.search(in_track.title):
+            in_track.flags.add_flag(TrackFlagCodes.p_intro)
+
+    def test_album_track_reprise(self, in_track: Track):
+        '''
+        Test the track to see if it is labeled as a
+        reprise, either to the album, or to a previous
+        track on the album.
+        '''
+        reprise_p = re.compile(r'[\s\W]?Reprise', re.IGNORECASE)
+        if reprise_p.search(in_track.title):
+            in_track.flags.add_flag(TrackFlagCodes.p_reprise)
+
+    def test_intro_reprise_presence(self, in_album: Album):
+        '''
+        Test to see if both an intro and reprise is
+        contained in the album, and add an additional
+        album level flag to identify it.
+        '''
+        reprise = False
+        intro = False
+        for dsc in sorted(in_album.discs):
+            for trk in sorted(in_album.discs[dsc].tracks,
+                              key=lambda x: x.track_no):
+                if TrackFlagCodes.p_intro in trk.flags:
+                    intro = True
+                if TrackFlagCodes.p_reprise in trk.flags:
+                    reprise = True
+                if intro and reprise:
+                    in_album.flags.add_flag(AlbumFlagCodes.p_intro_reprise)
+                    break